Refactor apriltag 3rdparty to migrate all tagAruco*.{c,h} files in aruco folder See PR https://github.com/AprilRobotics/apriltag/pull/436 Update also the README.VISP file that ecplains changes introduced in apriltag 3rdparty
