+++ 3rdparty/apriltag/CMakeLists.txt
@@ -11,8 +11,8 @@ set(APRILTAG_VERSION "${APRILTAG_VERSION_MAJOR}.${APRILTAG_VERSION_MINOR}.${APRI
 vp_include_directories(${CMAKE_CURRENT_SOURCE_DIR})
 
 file(GLOB tag_srcs tag*.c)
-file(GLOB lib_srcs *.c common/*.c)
-file(GLOB lib_hdrs *.h common/*.h common/sys/*.h)
+file(GLOB lib_srcs *.c aruco/*.c common/*.c)
+file(GLOB lib_hdrs *.h aruco/*.h common/*.h common/sys/*.h)
 
 vp_list_filterout(lib_srcs apriltag_pywrap.c)   # useless
 vp_list_filterout(lib_srcs common/pjpeg.c)      # useless
@@ -95,23 +95,23 @@ elseif(MSVC)
     vp_set_source_file_compile_flag(tag25h9.c /wd4996 /wd4244)
     vp_set_source_file_compile_flag(tag36h11.c /wd4996 /wd4244)
     vp_set_source_file_compile_flag(tagCircle21h7.c /wd4996 /wd4244)
-    vp_set_source_file_compile_flag(tagAruco4x4_50.c /wd4996 /wd4244)
-    vp_set_source_file_compile_flag(tagAruco4x4_100.c /wd4996 /wd4244)
-    vp_set_source_file_compile_flag(tagAruco4x4_250.c /wd4996 /wd4244)
-    vp_set_source_file_compile_flag(tagAruco4x4_1000.c /wd4996 /wd4244)
-    vp_set_source_file_compile_flag(tagAruco5x5_50.c /wd4996 /wd4244)
-    vp_set_source_file_compile_flag(tagAruco5x5_100.c /wd4996 /wd4244)
-    vp_set_source_file_compile_flag(tagAruco5x5_250.c /wd4996 /wd4244)
-    vp_set_source_file_compile_flag(tagAruco5x5_1000.c /wd4996 /wd4244)
-    vp_set_source_file_compile_flag(tagAruco6x6_50.c /wd4996 /wd4244)
-    vp_set_source_file_compile_flag(tagAruco6x6_100.c /wd4996 /wd4244)
-    vp_set_source_file_compile_flag(tagAruco6x6_250.c /wd4996 /wd4244)
-    vp_set_source_file_compile_flag(tagAruco6x6_1000.c /wd4996 /wd4244)
-    vp_set_source_file_compile_flag(tagAruco7x7_50.c /wd4996 /wd4244)
-    vp_set_source_file_compile_flag(tagAruco7x7_100.c /wd4996 /wd4244)
-    vp_set_source_file_compile_flag(tagAruco7x7_250.c /wd4996 /wd4244)
-    vp_set_source_file_compile_flag(tagAruco7x7_1000.c /wd4996 /wd4244)
-    vp_set_source_file_compile_flag(tagArucoMIP36h12.c /wd4996 /wd4244)
+    vp_set_source_file_compile_flag(aruco/tagAruco4x4_50.c /wd4996 /wd4244)
+    vp_set_source_file_compile_flag(aruco/tagAruco4x4_100.c /wd4996 /wd4244)
+    vp_set_source_file_compile_flag(aruco/tagAruco4x4_250.c /wd4996 /wd4244)
+    vp_set_source_file_compile_flag(aruco/tagAruco4x4_1000.c /wd4996 /wd4244)
+    vp_set_source_file_compile_flag(aruco/tagAruco5x5_50.c /wd4996 /wd4244)
+    vp_set_source_file_compile_flag(aruco/tagAruco5x5_100.c /wd4996 /wd4244)
+    vp_set_source_file_compile_flag(aruco/tagAruco5x5_250.c /wd4996 /wd4244)
+    vp_set_source_file_compile_flag(aruco/tagAruco5x5_1000.c /wd4996 /wd4244)
+    vp_set_source_file_compile_flag(aruco/tagAruco6x6_50.c /wd4996 /wd4244)
+    vp_set_source_file_compile_flag(aruco/tagAruco6x6_100.c /wd4996 /wd4244)
+    vp_set_source_file_compile_flag(aruco/tagAruco6x6_250.c /wd4996 /wd4244)
+    vp_set_source_file_compile_flag(aruco/tagAruco6x6_1000.c /wd4996 /wd4244)
+    vp_set_source_file_compile_flag(aruco/tagAruco7x7_50.c /wd4996 /wd4244)
+    vp_set_source_file_compile_flag(aruco/tagAruco7x7_100.c /wd4996 /wd4244)
+    vp_set_source_file_compile_flag(aruco/tagAruco7x7_250.c /wd4996 /wd4244)
+    vp_set_source_file_compile_flag(aruco/tagAruco7x7_1000.c /wd4996 /wd4244)
+    vp_set_source_file_compile_flag(aruco/tagArucoMIP36h12.c /wd4996 /wd4244)
     # disable optimization
     foreach(f ${tag_srcs})
       vp_set_source_file_compile_flag(${f} /O0)
