+++ 3rdparty/apriltag/aruco/tagAruco4x4_100.h
