+++ 3rdparty/apriltag/aruco/tagAruco4x4_1000.h
