+++ 3rdparty/apriltag/aruco/tagAruco4x4_50.h
