+++ 3rdparty/apriltag/aruco/tagAruco5x5_1000.c
