+++ 3rdparty/apriltag/aruco/tagAruco5x5_50.c
