+++ 3rdparty/apriltag/aruco/tagAruco6x6_100.h
