+++ 3rdparty/apriltag/aruco/tagAruco6x6_1000.c
