+++ 3rdparty/apriltag/aruco/tagAruco6x6_50.h
