+++ 3rdparty/apriltag/aruco/tagAruco7x7_1000.h
