+++ 3rdparty/apriltag/aruco/tagAruco7x7_50.c
